Majorly refactor how fonts are cached
Firstly, I added a FontProperties struct for organizing font properties,
now I won't have to worry about the ordering of weight and style when passing
the data to the font manager.

As an extension to the FontProperties struct, I created enums for FontWeight
and FontStyle instead of using Strings to store that info, since there are a
countable number of options those can correlate to. I also added a FontSize
struct just for easier code readability (in the future I should make a CursorX
and Y struct for code reading as well).

Next the reason I refactored the font caching mechanism by so much — (this
isn't ChatGPT writing this I just like the em dash here) every word needs to
either reload the font per word, or store a reference to the loaded font. The
reason is, words are processed one word at a time, and each word could be
rendered with a different font (by setting font-family attribute, em tag, i tag,
b tag, strong tag), and thus needs to have access to a new font per word.

My first solution was to just naively reload the font everytime with the old
font manager. This 'worked' but was so slow it was totally unfeasible. My goal
then was to be able to access a cached reference to the loaded ab_glyph and
RustyBuzz font objects. The issue is, they are only returned as references.

The solution that (with some help from ChatGPT) I arrived on was caching the
references, against the will of the Rust borrow checker. To get around the
borrow checker yelling at me, I used Box::leak() (thanks chat) to give the references a
static lifetime, so the fonts loaded into memory will not be dropped until
the program closes. But that's a sacrifice I'm willing to make.