Optimize font loading while creating display list
Fonts are only changed in the token function, thus, the font and
face can be loaded in token, and passed by reference into the word
function. This is slightly more efficient than reading from the
font manager hashmap over and over in the word function.

When the font properties are changed when a tag block is hit,
the font and face can be loaded from the font manager with the new
properties.